% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-utils.R
\name{.assert_output_quality}
\alias{.assert_output_quality}
\title{Assert output quality}
\usage{
.assert_output_quality(result, checks = list())
}
\arguments{
\item{result}{Result to check}

\item{checks}{List of checks to perform}
}
\description{
Performs a series of simple checks on a result object to ensure it does not
contain obviously problematic values.
}
\keyword{internal}
