% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_convolution.R
\name{.batch_convolution}
\alias{.batch_convolution}
\title{Batch convolution helper}
\usage{
.batch_convolution(signals, kernels, output_length)
}
\arguments{
\item{signals}{Matrix of stimulus regressors (time x regressors)}

\item{kernels}{Matrix of kernels/basis functions (time x kernels)}

\item{output_length}{Length of the desired output time series}
}
\description{
Performs convolution of multiple stimulus regressors with a set of
kernel columns and sums the results. Used by the parametric engine
to build design matrices efficiently.
}
\keyword{internal}
