% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock-solid-memory.R
\name{.cleanup_workspace}
\alias{.cleanup_workspace}
\title{Clean up temporary objects and force garbage collection}
\usage{
.cleanup_workspace(objects_to_remove = NULL, gc_runs = 2, verbose = FALSE)
}
\description{
Clean up temporary objects and force garbage collection
}
\keyword{internal}
