% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock-solid-recovery.R
\name{.construct_safe_default}
\alias{.construct_safe_default}
\title{Construct safe default values based on context}
\usage{
.construct_safe_default(context)
}
\description{
Construct safe default values based on context
}
\keyword{internal}
