% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{.diag_inform}
\alias{.diag_inform}
\title{Emit an informational message respecting package verbosity}
\usage{
.diag_inform(msg, ...)
}
\description{
Emit an informational message respecting package verbosity
}
\keyword{internal}
