% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{.diag_progressor}
\alias{.diag_progressor}
\title{Create a progressr progressor if available}
\usage{
.diag_progressor(steps)
}
\description{
Create a progressr progressor if available
}
\keyword{internal}
