% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf-interface-lwu.R
\name{.lwu_hrf_default_bounds}
\alias{.lwu_hrf_default_bounds}
\title{Internal LWU HRF default bounds}
\usage{
.lwu_hrf_default_bounds()
}
\value{
List with elements \code{lower} and \code{upper}
}
\description{
Internal LWU HRF default bounds
}
\keyword{internal}
