% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf-interface-lwu.R
\name{.lwu_hrf_default_seed}
\alias{.lwu_hrf_default_seed}
\title{Internal LWU HRF default seed}
\usage{
.lwu_hrf_default_seed()
}
\value{
Numeric vector of default parameter seed
}
\description{
Internal LWU HRF default seed
}
\keyword{internal}
