% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf-interface-lwu.R
\name{.lwu_hrf_parameter_names}
\alias{.lwu_hrf_parameter_names}
\title{Internal LWU HRF parameter names}
\usage{
.lwu_hrf_parameter_names()
}
\value{
Character vector of parameter names
}
\description{
Internal LWU HRF parameter names
}
\keyword{internal}
