% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-processing.R
\name{.parallel_gauss_newton}
\alias{.parallel_gauss_newton}
\title{Parallel Gauss-Newton optimization}
\usage{
.parallel_gauss_newton(
  hard_indices,
  fit_data,
  prepared_data,
  hrf_interface,
  parallel_config,
  theta_bounds,
  max_iter_gn = 5,
  tol_gn = 1e-04,
  lambda_ridge = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{hard_indices}{Indices of hard voxels}

\item{fit_data}{Current fit data}

\item{prepared_data}{Prepared input data}

\item{hrf_interface}{HRF interface functions}

\item{parallel_config}{Parallel configuration}

\item{...}{Additional GN parameters}
}
\value{
Updated fit results for hard voxels
}
\description{
Perform Gauss-Newton optimization for hard voxels in parallel.
}
\keyword{internal}
