% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock-solid-recovery.R
\name{.progressive_estimation}
\alias{.progressive_estimation}
\title{Progressive algorithm degradation}
\usage{
.progressive_estimation(
  Y_proj,
  S_target_proj,
  hrf_interface,
  theta_seed,
  theta_bounds,
  recenter_global_passes = 3,
  verbose = TRUE
)
}
\description{
Progressive algorithm degradation
}
\keyword{internal}
