% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock-solid-recovery.R
\name{.safe_combine_results}
\alias{.safe_combine_results}
\title{Safely combine voxel-wise results}
\usage{
.safe_combine_results(results, n_params)
}
\description{
Safely combine voxel-wise results
}
\keyword{internal}
