% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_performance_dispatcher.R
\name{.smart_parallel_decision}
\alias{.smart_parallel_decision}
\title{Smart parallel dispatcher}
\usage{
.smart_parallel_decision(n_voxels, n_cores = NULL, overhead_per_voxel = 0.001)
}
\arguments{
\item{n_voxels}{Number of voxels to process}

\item{n_cores}{Available cores}

\item{overhead_per_voxel}{Estimated overhead per voxel for parallel setup}
}
\value{
List with parallelization recommendation
}
\description{
Decides whether parallelization will be beneficial based on problem size
and available hardware resources.
}
