% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_performance_dispatcher.R
\name{.smart_qr_solve}
\alias{.smart_qr_solve}
\title{Smart QR solve dispatcher}
\usage{
.smart_qr_solve(X, Y, iteration_context = FALSE, lambda_ridge = 0)
}
\arguments{
\item{X}{Design matrix}

\item{Y}{Response matrix}

\item{iteration_context}{Are we in an iterative algorithm?}

\item{lambda_ridge}{Ridge penalty}
}
\value{
Solved coefficients
}
\description{
Chooses between cached and direct QR based on iteration context.
Provides caching benefits only when there are multiple iterations.
}
