% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engineering-standards.R
\name{.validate_constraints}
\alias{.validate_constraints}
\title{Validate numerical constraints}
\usage{
.validate_constraints(x, arg_name, constraints)
}
\description{
Validate numerical constraints
}
\keyword{internal}
