% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-hrf-fit-class.R
\name{get_fitted_values}
\alias{get_fitted_values}
\title{Get fitted values from parametric_hrf_fit}
\usage{
get_fitted_values(x, Y_proj = NULL)
}
\arguments{
\item{x}{parametric_hrf_fit object}

\item{Y_proj}{Original projected Y data (required if residuals not stored)}
}
\value{
Matrix of fitted values
}
\description{
Get fitted values from parametric_hrf_fit
}
\keyword{internal}
