% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-hrf-fit-class.R
\name{is_v2_fit}
\alias{is_v2_fit}
\title{Check if a parametric_hrf_fit has Sprint 2 enhancements}
\usage{
is_v2_fit(x)
}
\arguments{
\item{x}{parametric_hrf_fit object}
}
\value{
Logical indicating if object has Sprint 2 fields
}
\description{
Check if a parametric_hrf_fit has Sprint 2 enhancements
}
\keyword{internal}
