% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-hrf-fit-class.R
\name{n_timepoints}
\alias{n_timepoints}
\title{Number of time points used during fitting}
\usage{
n_timepoints(x)
}
\arguments{
\item{x}{parametric_hrf_fit}
}
\description{
Number of time points used during fitting
}
\keyword{internal}
