% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-hrf-fit-class.R
\name{n_voxels}
\alias{n_voxels}
\title{Number of voxels in a fit object}
\usage{
n_voxels(x)
}
\arguments{
\item{x}{parametric_hrf_fit}
}
\description{
Number of voxels in a fit object
}
\keyword{internal}
