% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-hrf-fit-methods.R
\name{residuals.parametric_hrf_fit}
\alias{residuals.parametric_hrf_fit}
\title{Extract residuals from a parametric_hrf_fit}
\usage{
\method{residuals}{parametric_hrf_fit}(object, ...)
}
\arguments{
\item{object}{A \code{parametric_hrf_fit} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Numeric matrix of residuals or NULL
}
\description{
Extract residuals from a parametric_hrf_fit
}
