% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engineering-standards.R
\name{set_engineering_options}
\alias{set_engineering_options}
\title{Set engineering standards options}
\usage{
set_engineering_options(
  verbose = FALSE,
  validate = TRUE,
  profile = FALSE,
  precision = "double",
  debug = FALSE
)
}
\description{
Set engineering standards options
}
