% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric-hrf-fit-methods.R
\name{vcov.parametric_hrf_fit}
\alias{vcov.parametric_hrf_fit}
\title{Variance-covariance matrix for parametric_hrf_fit parameters}
\usage{
\method{vcov}{parametric_hrf_fit}(object, voxel_index, ...)
}
\arguments{
\item{object}{A \code{parametric_hrf_fit} object}

\item{voxel_index}{Integer vector of voxel indices}

\item{...}{Currently unused}
}
\value{
A variance-covariance matrix or list of matrices
}
\description{
Returns the estimated variance-covariance matrix for the HRF parameters of
one or more voxels. If only a single voxel index is supplied a matrix is
returned; otherwise a list of matrices is produced. When full covariance
information was not retained during fitting, the matrix is constructed using
the available standard errors (i.e. a diagonal matrix).
}
