% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss-newton-refinement.R
\name{.get_jacobian_and_residuals}
\alias{.get_jacobian_and_residuals}
\title{Get Jacobian matrix and residuals for Gauss-Newton}
\usage{
.get_jacobian_and_residuals(theta, y, S, t_hrf, hrf_interface, n_time)
}
\description{
Get Jacobian matrix and residuals for Gauss-Newton
}
\keyword{internal}
