% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_linear_solver.R
\name{.ridge_linear_solve}
\alias{.ridge_linear_solve}
\title{Fast ridge regression solver using RcppEigen and OpenMP}
\usage{
.ridge_linear_solve(X, Y, lambda_ridge = 0)
}
\arguments{
\item{X}{Design matrix (time x parameters)}

\item{Y}{Response matrix (time x voxels)}

\item{lambda_ridge}{Ridge penalty}
}
\description{
Fast ridge regression solver using RcppEigen and OpenMP
\cr Error if non-finite values are detected in the inputs.
}
\keyword{internal}
