% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock-solid-validation.R
\name{.rock_solid_validate_inputs}
\alias{.rock_solid_validate_inputs}
\title{Master validation function that orchestrates all checks}
\usage{
.rock_solid_validate_inputs(
  fmri_data,
  event_model,
  parametric_hrf,
  theta_seed,
  theta_bounds,
  hrf_span,
  lambda_ridge,
  recenter_global_passes,
  recenter_epsilon,
  r2_threshold,
  mask,
  verbose,
  caller = "estimate_parametric_hrf"
)
}
\description{
Master validation function that orchestrates all checks
}
\keyword{internal}
