% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{.validate_event_model}
\alias{.validate_event_model}
\title{Validate event model}
\usage{
.validate_event_model(event_model, n_time, caller = "estimate_parametric_hrf")
}
\description{
Validate event model
}
\keyword{internal}
