% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-version.R
\docType{data}
\name{fmriparametric_api_version}
\alias{fmriparametric_api_version}
\title{Public API version}
\format{
An object of class \code{character} of length 1.
}
\usage{
fmriparametric_api_version
}
\description{
This string indicates the semantic version of the exported API
for compatibility checks.
}
\keyword{datasets}
